Crate mvpoly

This module contains the definition of the MVPoly trait, which is used to represent multi-variate polynomials.

Different representations are provided in the sub-modules:

monomials: a representation based on monomials

prime: a representation based on a mapping from variables to prime numbers. This representation is unmaintained for now. We leave it for interested users.

“Expressions”, as defined in the kimchi crate, can be converted into a multi-variate polynomial using the from_expr method.

Modules

monomials

pbt

This module contains a list of property tests for the MVPoly trait.

prime

Multivariate polynomial dense representation using prime numbers

utils

This module contains functions to work with prime numbers and to compute dimension of multivariate spaces

Traits

MVPoly

Generic trait to represent a multi-variate polynomial

Functions

compute_combined_cross_terms

Compute the cross terms of a list of polynomials. The polynomials are linearly combined using the power of a combiner, often called α.