mvpoly::monomials

Struct Sparse

pub struct Sparse<F: PrimeField, const N: usize, const D: usize> {
    pub monomials: HashMap<[usize; N], F>,
}

Represents a multivariate polynomial in N variables with coefficients in F. The polynomial is represented as a sparse polynomial, where each monomial is represented by a vector of N exponents.

For instance, the multivariate polynomial 3x^2 y + 2 x z will be represented by the HashMap:

[2, 1, 0] -> 3
 |  |  |
 x  y  z
 |  |  |
[1, 0, 1] -> 2

Fields
monomials: HashMap<[usize; N], F>