Struct mvpoly::prime::Dense

pub struct Dense<F: PrimeField, const N: usize, const D: usize> { /* private fields */ }

Represents a multivariate polynomial of degree less than D in N variables. The representation is dense, i.e., all coefficients are stored. The polynomial is represented as a vector of coefficients, where the index of the coefficient corresponds to the index of the monomial. A mapping between the index and the prime decomposition of the monomial is stored in normalized_indices.

Implementations

impl<F: PrimeField, const N: usize, const D: usize> Dense<F, N, D>

pub fn new() -> Self

pub fn iter(&self) -> impl Iterator<Item = &F>

pub unsafe fn random<RNG: RngCore>( rng: &mut RNG, max_degree: Option<usize> ) -> Self

Generate a random polynomial of maximum degree max_degree.

If None is provided as the maximum degree, the polynomial will be generated with a maximum degree of D.

Safety

Marked as unsafe to warn the user to use it with caution and to not necessarily rely on it for security/randomness in cryptographic protocols. The user is responsible for providing its own secure polynomial random generator, if needed.

For now, the function is only used for testing.

pub fn dimension() -> usize

pub fn from_coeffs(coeff: Vec<F>) -> Self

pub fn from_variable<C: Into<usize>>(var: C) -> Self

pub fn number_of_variables(&self) -> usize

pub fn maximum_degree(&self) -> usize

pub fn is_constant(&self) -> bool

pub unsafe fn degree(&self) -> usize

Returns the degree of the polynomial.

The degree of the polynomial is the maximum degree of the monomials that have a non-zero coefficient.

Safety

The zero polynomial as a degree equals to 0, as the degree of the constant polynomials. We do use the unsafe keyword to warn the user for this specific case.

pub fn compute_normalized_indices() -> Vec<usize>

Output example for N = 2 and D = 2:

- 0 -> 1
- 1 -> 2
- 2 -> 3
- 3 -> 4
- 4 -> 6
- 5 -> 9

pub fn is_homogeneous(&self) -> bool

Returns true if the polynomial is homoegenous of degree d

pub fn double(&self) -> Self

pub fn mul_by_scalar(&self, c: F) -> Self

pub fn eval(&self, x: &[F; N]) -> F

Evaluate the polynomial at the vector point x.

This is a dummy implementation. A cache can be used for the monomials to speed up the computation.

pub fn increase_degree<const D_PRIME: usize>(&self) -> Dense<F, N, D_PRIME>