Struct mvpoly::prime::Dense

pub struct Dense<F: PrimeField, const N: usize, const D: usize> { /* private fields */ }

Represents a multivariate polynomial of degree less than D in N variables. The representation is dense, i.e., all coefficients are stored. The polynomial is represented as a vector of coefficients, where the index of the coefficient corresponds to the index of the monomial. A mapping between the index and the prime decomposition of the monomial is stored in normalized_indices.

Implementations

impl<F: PrimeField, const N: usize, const D: usize> Dense<F, N, D>

pub fn new() -> Self

pub fn iter(&self) -> impl Iterator<Item = &F>

pub fn dimension() -> usize

pub fn from_coeffs(coeff: Vec<F>) -> Self

pub fn number_of_variables(&self) -> usize

pub fn maximum_degree(&self) -> usize

pub fn compute_normalized_indices() -> Vec<usize>

Output example for N = 2 and D = 2:

- 0 -> 1
- 1 -> 2
- 2 -> 3
- 3 -> 4
- 4 -> 6
- 5 -> 9

pub fn increase_degree<const D_PRIME: usize>(&self) -> Dense<F, N, D_PRIME>